AIOT_MQTT/AIOT_MQTT/requirements.txt
chainlit~=1.1.401
openai~=1.39.0
paho-mqtt~=2.1.0
